Split out MediaPage component and test
