+++ src/lib/components/regionSelect/regionSelect.svelte
@@ -38,7 +38,7 @@
 	disabled={isLoading}
 	selected={selectedRegion}
 >
-	<Select.Trigger {id} class="w-1/2 md:w-1/3 bg-gray-800 text-white border-gray-900">
+	<Select.Trigger {id} class="bg-gray-800 text-white border-gray-900">
 		<Select.Value placeholder="Region" />
 	</Select.Trigger>
 	<Select.Content class="bg-gray-800 text-white border-gray-900 overflow-y-scroll max-h-[30vh]">
